Allow only admin to create new rooms Allow only admin to destroy rooms

+++ app/controllers/rooms_controller.rb
@@ -17,19 +17,27 @@ class RoomsController < ApplicationController
   end
 
   def create
-    room = Room.new(room_params)
+    if current_user.role == 'admin'
+      room = Room.new(room_params)
 
-    if room.save
-      render json: room, status: :created
+      if room.save
+        render json: room, status: :created
+      else
+        render json: room.errors, status: :unprocessable_entity
+      end
     else
-      render json: room.errors, status: :unprocessable_entity
-    end
+      render json: 'Access not allowed !'
+    end    
   end
 
   def destroy
-    room = Room.find(params[:id])
-    room.destroy
-    head :no_content
+    if current_user.role == 'admin'
+      room = Room.find(params[:id])
+      room.destroy
+      head :no_content
+    else
+      render json: 'Access not allowed !'
+    end    
   end
 
   private
